requirements.txt
requests
python-dotenv
pydantic
python-dateutil
pytest